pyproject.toml
[project]
name = "rag-chat-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "google-genai>=1.73.1",
  "langchain>=1.2.15",
  "langchain-community>=0.4.1",
  "langchain-google-genai>=4.2.2",
  "langchain-postgres>=0.0.17",
  "psycopg2-binary>=2.9.12",
  "pypdf>=6.10.2",
  "python-dotenv>=1.2.2",
]
